add options,test utils and new tests
fix linting and add more test

fix casting and add document

fix  equal method and relevant tests

fix linting

the Correction of how the number of rows is calculated

fix linting
