+++ cpp/src/arrow/array/statistics.h
@@ -23,11 +23,13 @@
 #include <string>
 #include <variant>
 
+#include "arrow/compare.h"
+#include "arrow/result.h"
+#include "arrow/type_fwd.h"
 #include "arrow/util/visibility.h"
 
 namespace arrow {
-class DataType;
-struct Scalar;
+
 /// \class ArrayStatistics
 /// \brief Statistics for an Array
 ///
@@ -101,7 +103,13 @@ struct ARROW_EXPORT ArrayStatistics {
   bool is_max_exact = false;
 
   /// \brief Check two statistics for equality
-  bool Equals(const ArrayStatistics& other) const;
+  ///
+  /// \param[in] other the ArrayStatistics to compare with
+  /// \param[in] options the options for equality comparisons of Scalars
+  ///
+  /// \return true if  statistics are equal
+  bool Equals(const ArrayStatistics& other,
+              const EqualOptions& options = EqualOptions::Defaults()) const;
 
   /// \brief Check two statistics for equality
   bool operator==(const ArrayStatistics& other) const { return Equals(other); }
@@ -109,5 +117,34 @@ struct ARROW_EXPORT ArrayStatistics {
   /// \brief Check two statistics for not equality
   bool operator!=(const ArrayStatistics& other) const { return !Equals(other); }
 };
+
+namespace internal {
+/// \brief Extract all nested arrays from an array as \ref ArrayDataVector.
+///
+/// Returns all nested arrays within the given array, up to the specified
+/// maximum nesting depth, as a vector of ArrayData.
+///
+/// \param[in] array The input array from which nested arrays will be extracted.
+/// \param[in] max_nesting_depth The maximum depth of nested arrays to extract.
+///
+/// \return A vector of \ref ArrayData .
+ARROW_EXPORT
+Result<ArrayDataVector> ExtractColumnsToArrayData(const std::shared_ptr<Array>& array,
+                                                  const int32_t& max_nesting_depth);
+/// \brief Collect Statistics From ArrayDataVector and turns into an Array
+///
+/// \param[in] memory_pool the memory pool to allocate memory from
+/// \param extracted_array_data_vector The input ArrayData from which statistics array is
+/// extracted
+///
+/// \param num_rows The input specifies the number of rows an object if it  is set to
+/// std::nullopt, the number of rows is deduced from the length of the first array
+///
+/// \return A statistics array.
+ARROW_EXPORT
+Result<std::shared_ptr<Array>> MakeStatisticsArray(
+    MemoryPool* memory_pool, const ArrayDataVector& extracted_array_data_vector,
+    std::optional<int64_t> num_rows = std::nullopt);
+}  // namespace internal
 
 }  // namespace arrow
